Update README.md
In this Project. i build a Decision-Tree Classifier model to predict the safety of the car. build two models, one with criterion gini index and another one with criterion entropy. the model yeilds a very good performance as indicated by the model accuracy in both the casess which was found to be 0.527145.
In the model with criterion gini index , the training-set accuracy score is 0.786517 while the test-set accuracy to be 0.527145. these two values are quite comparable. so there is no sign of overfitting.
Smiliarly , in the model with criterion entropy , the training set accuracy score is 0.7865 while the test-set accuracy to be 0.527145. we get the same values as in the case with criterion gini. so there is no sign of overfitting.
In both the cases , the training set and test-set accuracy score is the same. it may happen because of small dataset.
The confusion matrix and classification report yeilds very good model performance.